Move TraitsOriginLabel to Components
